data_pipelines/stocks/services/alphavantage_loader/pyproject.toml
[project]
name = "alphavantage-loader"
version = "0.1.0"
description = "Alpha Vantage data BQ loader service"
requires-python = ">=3.11"
dependencies = [
    "alphavantage",
    "google-cloud-storage>=3.4.0",
    "google-cloud-bigquery>=3.38.0",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
alphavantage = { workspace = true }

# Optional: Add scripts if you want CLI commands
[project.scripts]
load = "alphavantage_loader.main:main"